Made changes to lower count of warnings

+++ src/kernel/memory/stack.c
@@ -2,11 +2,12 @@
 #include "drivers/uart/uart.h"
 #include "kernel/main.h"
 #include "include/board.h"
+#include "stack.h"
 
 #define stack_end_magic_num 0x22B3D20B07438DF4
 #define stack_top_magic_num 0x58C0E68DE152E781
 
-extern u8 _stack_top_test, _stack_bottom_test;
+extern u64 _stack_top_test, _stack_bottom_test;
 
 void stack_init() {
     *(volatile u64 *)(&_stack_bottom_test) = stack_end_magic_num;
@@ -19,9 +20,9 @@ void stack_init() {
 
 void stack_confirm() {
     if (*(volatile u64 *)(&_stack_bottom_test) != stack_end_magic_num) {
-        PANIC("DATA_HAS_BEEN_WRITTEN_BELOW_STACK", *(volatile u64 *)(&_stack_bottom_test), (s64)(&_stack_bottom_test), 0);
+        PANIC("DATA_HAS_BEEN_WRITTEN_BELOW_STACK", *(volatile u64 *)((u64)&_stack_bottom_test), (s64)(&_stack_bottom_test), 0);
     }
     if (*(volatile u64 *)(&_stack_top_test) != stack_top_magic_num) {
-        PANIC("DATA_HAS_BEEN_WRITTEN_ABOVE_STACK", *(volatile u64 *)(&_stack_top_test), (s64)(&_stack_top_test), 0);
+        PANIC("DATA_HAS_BEEN_WRITTEN_ABOVE_STACK", *(volatile u64 *)((u64)&_stack_top_test), (s64)(&_stack_top_test), 0);
     }
 }
